Extract output formatters for Platform, Extensions, Implements; move CliSchemaView to Views
- Create PlatformOutputFormatter with FormatFrameworks, FormatVersions, FormatAssemblies
- Create ExtensionsOutputFormatter with FormatResults
- Create ImplementsOutputFormatter with FormatResults
- Move CliSchemaView and CliSchemaContext from CliSchemaCommand.cs to Views/CliSchemaView.cs
- Remove 'using Markout' from PlatformCommand, ExtensionsCommand, ImplementsCommand
- Commands now delegate all Markout rendering to Output/ formatters
