+++ src/dotnet-inspect/Commands/CliSchemaCommand.cs
@@ -1,6 +1,7 @@
 using System.CommandLine;
 using System.CommandLine.Help;
 using DotnetInspector.Options;
+using DotnetInspector.Views;
 using Markout;
 
 namespace DotnetInspector.Commands;
@@ -104,28 +105,4 @@ public class CliSchemaCommand
 
         return name;
     }
-}
-
-[MarkoutSerializable(TitleProperty = nameof(Title), DescriptionProperty = nameof(Description))]
-public class CliSchemaView
-{
-    [MarkoutIgnore]
-    public string Name { get; set; } = "";
-
-    [MarkoutIgnore]
-    public string Version { get; set; } = "";
-
-    [MarkoutIgnore]
-    public string Title => $"{Name} {Version}";
-
-    [MarkoutIgnore]
-    public string Description { get; set; } = "";
-
-    [MarkoutIgnoreInTable]
-    public List<TreeNode> Commands { get; set; } = [];
-}
-
-[MarkoutContext(typeof(CliSchemaView))]
-public partial class CliSchemaContext : MarkoutSerializerContext
-{
 }
