+++ src/dotnet-inspect/Commands/ImplementsCommand.cs
@@ -5,7 +5,6 @@ using DotnetInspector.Inspectors;
 using DotnetInspector.Metadata;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
-using Markout;
 
 namespace DotnetInspector.Commands;
 
@@ -120,37 +119,7 @@ public class ImplementsCommand
 
     private static void WriteMarkoutOutput(string targetType, List<ImplementerResult> results)
     {
-        var writer = new MarkoutWriter();
-        
-        writer.WriteHeading(1, $"Types Implementing {targetType}");
-
-        if (results.Count == 0)
-        {
-            writer.WriteParagraph("No implementing types found.");
-            Console.WriteLine(writer.ToString());
-            return;
-        }
-
-        writer.WriteField("Matches", results.Count);
-
-        // Group by source
-        var bySource = results.GroupBy(r => r.Source).ToList();
-
-        foreach (var sourceGroup in bySource)
-        {
-            var source = sourceGroup.Key;
-            var version = sourceGroup.First().SourceVersion;
-            var sourceDisplay = version != null ? $"{source}@{version}" : source;
-
-            writer.WriteHeading(2, sourceDisplay ?? "Unknown");
-
-            var headers = new[] { "Type", "Kind", "Relationship", "Assembly" };
-            var rows = sourceGroup.OrderBy(r => r.TypeName)
-                .Select(impl => new[] { impl.TypeName, impl.Kind, impl.Relationship, impl.Assembly ?? "" });
-            writer.WriteTable(headers, rows);
-        }
-
-        Console.WriteLine(writer.ToString());
+        Console.WriteLine(ImplementsOutputFormatter.FormatResults(targetType, results));
     }
 }
