+++ src/dotnet-inspect/Commands/PlatformCommand.cs
@@ -3,7 +3,6 @@ using DotnetInspector.Inspectors;
 using DotnetInspector.Metadata;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
-using Markout;
 
 namespace DotnetInspector.Commands;
 
@@ -64,34 +63,7 @@ public class PlatformCommand
             return 0;
         }
 
-        var writer = new MarkoutWriter();
-        writer.WriteHeading(1, "Installed Frameworks");
-
-        // Show summary info in detailed mode
-        if (options.Verbosity == Verbosity.Detailed)
-        {
-            var latestVersion = frameworks.Max(f => f.LatestVersion) ?? "";
-            var majorVersion = latestVersion.Contains('.') ? latestVersion[..latestVersion.IndexOf('.')] + ".0" : latestVersion;
-            var majorVersions = frameworks
-                .SelectMany(f => f.AllVersions)
-                .Select(v => v.Contains('.') ? v[..v.IndexOf('.')] : v)
-                .Distinct()
-                .Count();
-            var dotnetRoot = Path.GetDirectoryName(packsDir) ?? packsDir;
-
-            writer.WriteCompactFields(
-                new MarkoutField("Latest", majorVersion),
-                new MarkoutField("Patch", latestVersion),
-                new MarkoutField("Majors", majorVersions.ToString()),
-                new MarkoutField("Runtimes", frameworks.Count.ToString()),
-                new MarkoutField("Location", dotnetRoot));
-        }
-
-        var headers = new[] { "Framework", "Version", "Assemblies" };
-        var rows = frameworks.Select(f => new[] { f.ShortName, f.LatestVersion, f.AssemblyCount.ToString() });
-        writer.WriteTable(headers, rows);
-
-        Console.WriteLine(writer.ToString());
+        Console.WriteLine(PlatformOutputFormatter.FormatFrameworks(frameworks, options.Verbosity, packsDir));
         return 0;
     }
 
@@ -132,28 +104,7 @@ public class PlatformCommand
             return 0;
         }
 
-        var writer = new MarkoutWriter();
-        writer.WriteHeading(1, "Installed Versions");
-
-        foreach (var framework in frameworks)
-        {
-            writer.WriteHeading(2, framework.ShortName);
-            
-            var versions = framework.AllVersions;
-            if (options.Limit.HasValue)
-            {
-                versions = versions.Take(options.Limit.Value).ToList();
-            }
-
-            writer.WriteArray(versions);
-
-            if (options.Limit.HasValue && framework.AllVersions.Count > options.Limit.Value)
-            {
-                writer.WriteParagraph($"... *and {framework.AllVersions.Count - options.Limit.Value} more*");
-            }
-        }
-
-        Console.WriteLine(writer.ToString());
+        Console.WriteLine(PlatformOutputFormatter.FormatVersions(frameworks, options.Limit));
         return 0;
     }
 
@@ -192,56 +143,10 @@ public class PlatformCommand
             return ListAssembliesJson(frameworks, requestedVersion, options);
         }
 
-        var writer = new MarkoutWriter();
-
-        // Add header if multiple frameworks
-        if (frameworks.Count > 1 || string.IsNullOrEmpty(options.Framework))
-        {
-            writer.WriteHeading(1, "Platform Assemblies");
-            writer.WriteField("Packs Directory", packsDir);
-        }
-
-        foreach (var framework in frameworks)
-        {
-            var version = requestedVersion ?? framework.LatestVersion;
-            var refPath = PlatformResolver.GetRefAssemblyPath(framework.Path, version);
-            
-            if (refPath == null)
-            {
-                logger.Log($"Could not find ref path for {framework.ShortName} {version}");
-                continue;
-            }
-
-            var assemblies = PlatformResolver.GetAssemblies(refPath);
-
-            writer.WriteHeading(2, $"{framework.ShortName} ({version})");
-
-            var displayAssemblies = assemblies.AsEnumerable();
-            if (options.Limit.HasValue)
-            {
-                displayAssemblies = displayAssemblies.Take(options.Limit.Value);
-            }
-
-            if (options.IncludeTypes)
-            {
-                var headers = new[] { "Assembly", "Types" };
-                var rows = displayAssemblies.Select(a => new[] { a.Name, CountPublicTypes(a.Path).ToString() });
-                writer.WriteTable(headers, rows);
-            }
-            else
-            {
-                var headers = new[] { "Assembly" };
-                var rows = displayAssemblies.Select(a => new[] { a.Name });
-                writer.WriteTable(headers, rows);
-            }
-
-            if (options.Limit.HasValue && assemblies.Count > options.Limit.Value)
-            {
-                writer.WriteParagraph($"... *and {assemblies.Count - options.Limit.Value} more assemblies*");
-            }
-        }
-
-        Console.WriteLine(writer.ToString());
+        var multipleFrameworks = frameworks.Count > 1 || string.IsNullOrEmpty(options.Framework);
+        Console.WriteLine(PlatformOutputFormatter.FormatAssemblies(
+            frameworks, requestedVersion, options.IncludeTypes, options.Limit,
+            packsDir, multipleFrameworks, CountPublicTypes, logger));
         return 0;
     }
